requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
PyJWT==2.9.0
bcrypt==4.2.1
pydantic[email]==2.10.3
python-dotenv==1.0.1
